doc: add README.md

+++ services/abusive-user-checker/pkg/services/checker.go
@@ -39,7 +39,7 @@ func (*CheckerServiceServer) CheckImage(ctx context.Context, req *protos.CheckIm
 	}
 
 	return &protos.CheckImageResponse{
-		Id:            checked.Id,
+		Id:            int32(checked.Id),
 		PhashDistance: int32(checked.HammingDistance),
 		Hashes: &protos.ImageHashes{
 			Md5:    checked.MD5,
